+++ src/api2/managers/index.js
@@ -1,0 +1,3 @@
+export { FactorsManager } from './factors-manager'
+export { SignersManager } from './signers-manager'
+export { WalletsManager } from './wallets-manager'