+++ src/base/operations/cancel_atomic_swap_bid_builder.spec.js
@@ -5,13 +5,13 @@ import { default as xdr } from '../generated/xdr_generated'
 describe('cancel atomic swap bid ', function () {
   it('Success', function () {
     let bidID = '1408'
-    let op = CancelAtomicSwapBidBuilder.cancelASwapBid({
+    let op = CancelAtomicSwapBidBuilder.cancelAtomicSwapBid({
       bidID
     })
     let xdrOp = op.toXDR('hex')
     let operation = xdr.Operation.fromXDR(Buffer.from(xdrOp, 'hex'))
     let obj = Operation.operationToObject(operation)
-    expect(obj.type).to.be.equal('cancelAswapBid')
+    expect(obj.type).to.be.equal('cancelAtomicSwapBid')
     expect(obj.bidID).to.be.equal(bidID)
   })
 })