+++ src/base/operations/create_atomic_swap_request_builder.spec.js
@@ -11,11 +11,11 @@ describe('Create ASwapBidCreation request', function () {
       quoteAsset: 'ETH',
       creatorDetails: { data: 'new atomic swap' }
     }
-    let op = CreateAtomicSwapRequestBuilder.createASwapRequest(opts)
+    let op = CreateAtomicSwapRequestBuilder.createAtomicSwapRequest(opts)
     let xdrOp = op.toXDR('hex')
     let operation = xdr.Operation.fromXDR(Buffer.from(xdrOp, 'hex'))
     let obj = Operation.operationToObject(operation)
-    expect(obj.type).to.be.equal(xdr.OperationType.createAswapRequest().name)
+    expect(obj.type).to.be.equal(xdr.OperationType.createAtomicSwapRequest().name)
     expect(obj.baseAmount).to.be.equal(opts.baseAmount)
     expect(obj.bidID).to.be.equal(opts.bidID)
     expect(obj.quoteAsset).to.be.equal(opts.quoteAsset)