+++ test/helpers/index.js
@@ -19,6 +19,8 @@ import { StampHelper } from './stamp'
 import { LicenseHelper } from './license'
 import { AtomicSwapBid } from './atomic_swap_bid'
 import { AtomicSwap } from './atomic_swap'
+import { Poll } from './poll'
+import { Vote } from './vote'
 
 export const sdk = new TokenD(config.api_url, {
   allowHttp: config.allow_http
@@ -51,3 +53,5 @@ export const stampHelper = new StampHelper(DEFAULTS)
 export const licenseHelper = new LicenseHelper(DEFAULTS)
 export const atomicSwapBidHelper = new AtomicSwapBid(DEFAULTS)
 export const atomicSwapHelper = new AtomicSwap(DEFAULTS)
+export const pollHelper = new Poll(DEFAULTS)
+export const voteHelper = new Vote(DEFAULTS)