+++ CHANGELOG.md
@@ -17,23 +17,127 @@ for further information about branching and tagging conventions.
 
 - operation builders related to atomic swap
 
-### Changed
+## [1.7.0-x.2] - 2019-05-15
+### Added
+- Method `useBaseURL` to api-caller
+- Method `useNetworkDetails` to api-caller
+- Method `withWallet` to api-caller
+- Method `useApi` to factors-manager
+- Method `useApi` to wallet-manager
 
-- XDRs, commit 07643f13fd5610cc258999bbcb8dc931fba0f84d
+## [1.7.0-x.1] - 2019-05-07
+### Added
+- New managers classes for encapsulating actions with wallets,
+  factors, and change signers operations:
+  - `WalletsManager`
+  - `FactorsManager`
+  - `SignersManager`
+
+### Changed
+- Now using default axios instance in `ServerErrorBase` constructor
+  if there are no passed one
+
+## [1.7.0-x.0] - 2019-05-07
+### Added
+- Manage account specific rule builder
+- Release sanity check script
+
+### Changed
+- Use XDR version 3.3.0-x.0
+
+## [1.6.0] - 2019-05-07
+
+## [1.6.0-rc.0] - 2019-05-02
+### Added
+- Builders for `CancelPollOp`, `UpdatePollEndTime`
+- Unit tests for for `CancelPollOp`, `UpdatePollEndTime`
+- Integration tests for `CancelPollOp`, `UpdatePollEndTime`
+
+### Changed
+- Use XDR version 3.2.0
+
+## [1.6.0-x.0] - 2019-04-26
+### Added
+- `DELETE` method for ApiCaller
+
+### Fixed
+- Parsing transaction errors without operation details
+- Wallet verification request payload
+
+## [1.5.0] - 2019-04-19
+
+## [1.5.0-rc.1] - 2019-04-19
+### Added
+- `getRaw` method & `networkDetails` getter to API caller
+
+## [1.5.0-rc.0] - 2019-04-19
+### Added
+- `BLOB_TYPES` constant
+
+## [1.4.4] - 2019-04-15
+### Fixed
+- Param `referrerId` in create wallet POST request
+
+## [1.4.3] - 2019-04-08
+### Added
+- Added `errorResults` getter for `TransactionError` to simplify deriving operation error codes and messages
+
+## [1.4.2] - 2019-04-05
+### Fixed
+- Add missing `referrerId` arg to `wallets.create()`
+
+## [1.4.1] - 2019-04-04
+### Fixed
+- Repository link in package.json
+
+## [1.4.0] - 2019-04-03
+### Fixed
+- Use XDR version 3.1.1 (add pollID in result of create `CREATE_POLL` request)
+
+## [1.4.0-rc.0] - 2019-03-30
+### Added
+- Voting:
+  - ManageCreatePollRequestBuilder
+  - ManagePollBuilder
+  - ManageVoteBuilder
+  - Integration tests for voting
+
+### Changed
+- Use XDR version 3.1.0
+
+## [1.3.1-x.2] - 2019-03-21
+### Added
+- New "bravo" blob type
 
 ## [1.3.1-x.1] - 2019-03-14
-### Fixed
-- Use deserialize cache stub when creating a `Jsona` instance
-- Integration tests
-
 ### Added
 - Error codes for set fee op and op related to limits https://github.com/tokend/xdr/pull/43/files
 
 ### Changed
 - Use XDR version 3.0.1-x.0
 
+### Fixed
+- Use deserialize cache stub when creating a `Jsona` instance
+- Integration tests
+
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.3.1-x.1...HEAD
+[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.7.0-x.2...HEAD
+[1.7.0-x.2]: https://github.com/tokend/new-js-sdk/compare/1.7.0-x.1...1.7.0-x.2
+[1.7.0-x.1]: https://github.com/tokend/new-js-sdk/compare/1.7.0-x.0...1.7.0-x.1
+[1.7.0-x.0]: https://github.com/tokend/new-js-sdk/compare/1.6.0...1.7.0-x.0
+[1.6.0]: https://github.com/tokend/new-js-sdk/compare/1.6.0-rc.0...1.6.0
+[1.6.0-rc.0]: https://github.com/tokend/new-js-sdk/compare/1.6.0-x.0...1.6.0-rc.0
+[1.6.0-x.0]: https://github.com/tokend/new-js-sdk/compare/1.5.0...1.6.0-x.0
+[1.5.0]: https://github.com/tokend/new-js-sdk/compare/1.5.0-rc.1...1.5.0
+[1.5.0-rc.1]: https://github.com/tokend/new-js-sdk/compare/1.5.0-rc.0...1.5.0-rc.1
+[1.5.0-rc.0]: https://github.com/tokend/new-js-sdk/compare/1.4.4...1.5.0-rc.0
+[1.4.4]: https://github.com/tokend/new-js-sdk/compare/1.4.3...1.4.4
+[1.4.3]: https://github.com/tokend/new-js-sdk/compare/1.4.2...1.4.3
+[1.4.2]: https://github.com/tokend/new-js-sdk/compare/1.4.1...1.4.2
+[1.4.1]: https://github.com/tokend/new-js-sdk/compare/1.4.0...1.4.1
+[1.4.0]: https://github.com/tokend/new-js-sdk/compare/1.4.0-rc.0...1.4.0
+[1.4.0-rc.0]: https://github.com/tokend/new-js-sdk/compare/1.3.1-x.2...1.4.0-rc.0
+[1.3.1-x.2]: https://github.com/tokend/new-js-sdk/compare/1.3.1-x.1...1.3.1-x.2
 [1.3.1-x.1]: https://github.com/tokend/new-js-sdk/compare/1.3.0...1.3.1-x.1
 [1.3.0]: https://github.com/tokend/new-js-sdk/releases/tag/1.3.0